+++ cursus/push_swap/src/sort_utils.c
@@ -6,20 +6,24 @@
 /*   By: sergio-jimenez <sergio-jimenez@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/07 18:17:00 by sergio-jime       #+#    #+#             */
-/*   Updated: 2025/06/16 16:41:11 by sergio-jime      ###   ########.fr       */
+/*   Updated: 2025/06/17 11:33:21 by sergio-jime      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
 /**
- * calcula el valor minimo del stack y le asigna un indice
+ * @brief Finds the position (index) of the minimum value in the stack.
+ * @param stack Double pointer to the head of the stack.
+ * @return The index position (0-based) of the minimum value in the stack.
+ * @note Returns EXIT_FAILURE if stack is NULL or empty.
+ * @note The position refers to the node's sequential order (0 = head).
  */
-size_t	min_stack_value(t_stack **stack)
+int	min_stack_value(t_stack **stack)
 {
 	int		min;
 	int		i;
-	size_t	index;
+	int		index;
 	t_stack	*temp;
 
 	if (!stack || !(*stack))
@@ -33,7 +37,7 @@ size_t	min_stack_value(t_stack **stack)
 		if (temp->value < min)
 		{
 			min = temp->value;
-			index = (size_t)i;
+			index = i;
 		}
 		temp = temp->next;
 		i++;
@@ -42,7 +46,11 @@ size_t	min_stack_value(t_stack **stack)
 }
 
 /**
- * rb count
+ * @brief Counts the number of rotations needed to bring a target index
+ * to the top.
+ * @param stack Double pointer to the head of the stack.
+ * @param index Target index to locate.
+ * @return Number of rotations (rb) needed, or -1 if index not found.
  */
 int	rb_count(t_stack **stack, int index)
 {
@@ -62,7 +70,9 @@ int	rb_count(t_stack **stack, int index)
 }
 
 /**
- * calcula el mayor indice del stack.
+ * @brief Finds the maximum index value present in the stack.
+ * @param stack Double pointer to the head of the stack.
+ * @return The highest index value found in the stack.
  */
 int	max_stack_index(t_stack **stack)
 {
@@ -81,9 +91,9 @@ int	max_stack_index(t_stack **stack)
 }
 
 /**
- * @brief Calculate the size of the stack.
- * @param stack Pointer to the head of the stack.
- * @return The size of the stack.
+ * @brief Calculate the number of elements in the stack.
+ * @param stack Double pointer to the head of the stack.
+ * @return The total number of elements in the stack.
  */
 int	stack_size(t_stack **stack)
 {
@@ -100,6 +110,17 @@ int	stack_size(t_stack **stack)
 	return (size);
 }
 
+/**
+ * @brief Prints detailed information about each node in the stack.
+ * Prints for each node:
+ * - Index value.
+ * - Numeric value.
+ * - Memory address of the next node.
+ * @param head Pointer to the head node of the stack.
+ * @note Outputs "NULL List" if stack is empty.
+ * @note Debug function to explain the algorithm, content and structures.
+ * @note The output format is designed for debugging linked list integrity.
+ */
 void	ft_ps_printlist(t_stack *head)
 {
 	t_stack	*temp;
